+++ NEWS.md
@@ -1,9 +1,13 @@
-# happign 1.0.0
-* `get_wms_raster` the door now uses gdalwarp from `sf::gdal_utils` to avoid 
+# happign 0.2.0
+* `get_wms_raster` now uses gdalwarp from `sf::gdal_utils` to avoid 
 calculation and downloading of several tiles
 * update cog_2022 dataset to cog_2023
 * remove `get_apicarto_commune()` now supported by `get_apicarto_cadastre()`
 * `jsonlite` is used instead of `geojsonsf` package
+* Some `get_wms_raster` and `get_wfs` parameter names have been changed :
+   - shape -> x
+   - layer_name -> layer
+   - resolution -> res
 
 # happign 0.1.9
 * add `get_apicarto_viticole()`